moved domready logic into its own module
