+++ app/js/app/modules/details-toggle.js
